// react-module/src/openapi/models/warehouse-action-extended-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface WarehouseActionExtendedDto
 */
export interface WarehouseActionExtendedDto {
    /**
     * 
     * @type {number}
     * @memberof WarehouseActionExtendedDto
     */
    id?: any;
    /**
     * 
     * @type {Date}
     * @memberof WarehouseActionExtendedDto
     */
    createdAt?: any;
    /**
     * 
     * @type {string}
     * @memberof WarehouseActionExtendedDto
     */
    type?: WarehouseActionExtendedDtoTypeEnum;
    /**
     * 
     * @type {StorageAreaBasicDto}
     * @memberof WarehouseActionExtendedDto
     */
    sourceArea?: any;
    /**
     * 
     * @type {StorageAreaBasicDto}
     * @memberof WarehouseActionExtendedDto
     */
    targetArea?: any;
    /**
     * 
     * @type {AtomicProductDefinitionBasicDto}
     * @memberof WarehouseActionExtendedDto
     */
    productDefinition?: any;
    /**
     * 
     * @type {number}
     * @memberof WarehouseActionExtendedDto
     */
    amount?: any;
    /**
     * 
     * @type {string}
     * @memberof WarehouseActionExtendedDto
     */
    user?: any;
}

/**
    * @export
    * @enum {string}
    */
export enum WarehouseActionExtendedDtoTypeEnum {
    ADDITION = 'ADDITION',
    REMOVAL = 'REMOVAL',
    DISPLACEMENT = 'DISPLACEMENT',
    RESERVATION = 'RESERVATION',
    RESERVATIONCANCELLATION = 'RESERVATION_CANCELLATION'
}